// src/ROL-ECOyOP_Ant/components/CargaNuevoRol/interfaces.ts
export interface Err {
    msg: string,
    desc?: string[],
    canNotDismiss?: boolean, 
}

export interface Err_page {
    hoja: string,
    err: Err[] | null
}


export type HeaderData = {
    i_ColsNames:        number, 
    pag:                number, 
    hojaName:           string, 
    data:               HeaderRol, 
    op1SacaDiasPares:   boolean, 
    diasFestivos?:       { 
        existLabel: string, 
        lista: {
            df_dias: string, 
            df_servicios: string

        }[]
    }, 
}
// export type HeaderData = {
//     indexColsNames: number,
//     isPar:          Boolean, 
//     df_dias:        string,     //? ???
//     df_servicios:   string,     //? '1, 3, 4, 5, 8'
//     df_fila:        string      //? fila que contiene la celda con el mensaje 'Dias Festivos'
// }

export interface HeaderRol {
    "periodo_date_fin"?: Date | string,
    "periodo_date_ini"?: Date | string,
    "periodo"?      : string,
    "periodo_id"?   : number,
    "ori_des"?      : string,
    "ruta"?         : string | import("read-excel-file/types").Cell,
    "ruta_id"?      : number,
    "modalidad"?    : string,
    "modalidad_id"? : number,
}